requirements.txt
jupyter>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
folium>=0.12.0
ipython>=7.0.0
google-genai
tqdm
Pillow